+++ src/Constants.php
@@ -1,0 +1,19 @@
+<?php
+namespace samizdam\Geometry;
+
+/**
+ * You can define and you own class named G
+ */
+
+// @codeCoverageIgnoreStart
+if (! class_exists(Constants::class, true)) {
+
+    /**
+     *
+     * @author samizdam
+     */
+    abstract class Constants extends DefaultConstants
+    {
+    }
+}
+// @codeCoverageIgnoreEnd