+++ src/DefaultConstants.php
@@ -1,0 +1,17 @@
+<?php
+namespace samizdam\Geometry;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+abstract class DefaultConstants
+{
+
+    /**
+     *
+     * @var unknown
+     */
+    const π = M_PI;
+}