+++ src/Plane/Angle/AngleInterface.php
@@ -1,0 +1,10 @@
+<?php
+namespace samizdam\Geometry\Plane\Angle;
+
+interface AngleInterface
+{
+
+    public function getVertexPoint();
+
+    public function getSize();
+}