+++ src/Plane/Angle/AngleSizeUnitsEnum.php
@@ -1,0 +1,10 @@
+<?php
+namespace samizdam\Geometry\Plane\Angle;
+
+interface AngleSizeUnitsEnum
+{
+
+    const DEG = 'degree';
+
+    const RAD = 'radian';
+}